Add a display filter
You can now set a CPPredicate that will be use to filter the actual displayed content
This will not change the actual data source content. The advantage is that you can
keep objects in the array you need, but you don't want to display.

for instance

```objj
[myDataSource setDisplayFilter:[CPPredicate predicateWithFormat:@"not name like %@", aName]];
```

object with name like aName will still be accessible in [myDataSource content], but will
not be displayed in table view.