Transparent support for filter as predicate
TNTableViewDataSource now supports filtering using predicate.
The old method is still working, but if the input string (from the CPSearchField) can be parsed as a predicate,
the filtering will be done using that predicaye.

For instance you can type:

"john" -> it will filter all object with a keypath defined in searchableKeyPaths mathing john
"firstName contains 'john' and age > 10" -> filter all the John's are more than 10

You can also set the filter using setFilter:(id)aFilter. aFilter can be a CPString or a CPPredicate